% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{method_URL_table}
\alias{method_URL_table}
\title{Relevent URLs of methods}
\format{
A tibble:
\describe{
  \item{\code{short_name}}{The short name of simulation methods}
  \item{\code{URL}}{The URLs of method}
  \item{\code{Vignette}}{The vignette URLs of methods}
}
}
\usage{
method_URL_table
}
\description{
A tibble containing the URLs of methods.
}
\keyword{datasets}
