% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models}
\alias{models}
\title{Prediction Models of Each Method}
\format{
An object of class \code{list} of length 4.
}
\usage{
models
}
\description{
RF models for predicting time and memory with a given number of cell and gene in both parameter estimation and simulation.
}
\keyword{datasets}
