% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_memory_prediction.R
\name{prediction}
\alias{prediction}
\title{Predict Time and Memory}
\usage{
prediction(model, cell_num, gene_num)
}
\arguments{
\item{model}{Established RF model}

\item{cell_num}{The number of cell that need to simulated}

\item{gene_num}{The number of gene that need to simulated}
}
\description{
Predict Time and Memory
}
