% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plot.R
\name{render_funkyheatmap}
\alias{render_funkyheatmap}
\title{Render Funkyheatmap}
\usage{
render_funkyheatmap(data)
}
\arguments{
\item{data}{The overall data of this benchmarking study}
}
\description{
Render Funkyheatmap
}
